requirements.txt
django==2.2.1
azure_storage==0.36.0
djangorestframework==3.11.0
azure-cosmos==3.1.2
whitenoise == 5.0.1
django-oauth-toolkit == 1.3.2
mysqlclient == 1.4.6
django-suit==0.2.28
